+++ PokeTactics.Api/ServiceCollectionExtensions.cs
@@ -1,5 +1,7 @@
 using PokeTactics.Api.HostedServices;
+using PokeTactics.Api.SyncServices;
 using PokeTactics.Api.Utils;
+using PokeTactics.Core.Interfaces.SyncServices;
 using PokeTactics.Core.Utils.Extensions;
 
 namespace PokeTactics.Api;
@@ -10,6 +12,7 @@ public static class ServiceCollectionExtensions
     {
         services.AddExternalApiHttpClient();
         services.AddHostedServices();
+        services.AddSyncServices();
     }
 
     private static void AddExternalApiHttpClient(this IServiceCollection services)
@@ -23,5 +26,12 @@ public static class ServiceCollectionExtensions
     private static void AddHostedServices(this IServiceCollection services)
     {
         services.AddHostedService<PokemonSyncHostedService>();
+    }
+
+    private static void AddSyncServices(this IServiceCollection services)
+    {
+        services.AddScoped<IAbilitySyncService, AbilitySyncService>();
+        services.AddScoped<IMoveSyncService, MoveSyncService>();
+        services.AddScoped<IPokemonSyncService,PokemonSyncService>();
     }
 }
