+++ PokeTactics.Core/Interfaces/SyncServices/IAbilitySyncService.cs
@@ -1,0 +1,6 @@
+namespace PokeTactics.Core.Interfaces.SyncServices;
+
+public interface IAbilitySyncService
+{
+    Task Sync(CancellationToken cancellationToken);
+}
