+++ PokeTactics.Core/Interfaces/SyncServices/IMoveSyncService.cs
@@ -1,0 +1,6 @@
+namespace PokeTactics.Core.Interfaces.SyncServices;
+
+public interface IMoveSyncService
+{
+    Task Sync(CancellationToken cancellationToken);
+}
