+++ PokeTactics.Core/Interfaces/SyncServices/IPokemonSyncService.cs
@@ -1,0 +1,6 @@
+namespace PokeTactics.Core.Interfaces.SyncServices;
+
+public interface IPokemonSyncService
+{
+    Task Sync(CancellationToken cancellationToken);
+}
